requirements.txt
# Django and API Framework
Django>=5.0.0
django-ninja>=1.1.0
django-cors-headers>=4.3.0
django-extensions>=3.2.0

# Database and Caching
psycopg2-binary>=2.9.7
redis>=5.0.0
django-redis>=5.4.0

# Background Tasks and WebSockets
celery>=5.3.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0
channels>=4.2.0
channels-redis>=4.2.0

# AI/ML Core Libraries
torch>=2.1.0
transformers>=4.40.0
datasets>=2.15.0
accelerate>=0.26.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
numba>=0.57.0
llvmlite>=0.40.0

# Quantization support (optional - not available on all platforms like Apple Silicon)
# Install manually if needed: uv pip install bitsandbytes
# bitsandbytes>=0.43.0

# Medical/Scientific NLP (optional - install separately if needed)
# spacy>=3.6.1,<3.7.0
# scispacy>=0.5.3,<=0.5.4

# Hyperparameter Optimization
optuna>=3.4.0
optuna-dashboard>=0.13.0

# Data Processing and Analysis
polars>=0.20.0
numpy>=1.25.0

# Visualization
plotly>=5.17.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
matplotlib>=3.8.0
seaborn>=0.12.0

# File Processing
openpyxl>=3.1.0
xlrd>=2.0.1

# Model Interpretability
shap>=0.48.0

# Utilities
python-decouple>=3.8
Pillow>=10.0.0
python-magic>=0.4.27
tqdm>=4.66.0

# Development and Testing
pytest>=7.4.0
pytest-django>=4.6.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
flake8>=6.1.0
black>=23.9.0

# Monitoring and Logging
sentry-sdk>=1.38.0

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0
